Fixed missing sizes on slabs and others
