+++ src/main/java/net/dries007/tfc/objects/blocks/BlocksTFC.java
@@ -32,8 +32,8 @@ import net.dries007.tfc.objects.blocks.stone.BlockRockVariant;
 import net.dries007.tfc.objects.blocks.stone.BlockWallTFC;
 import net.dries007.tfc.objects.blocks.wood.*;
 import net.dries007.tfc.objects.fluids.FluidsTFC;
-import net.dries007.tfc.objects.items.itemblock.ItemBlockTFC;
-import net.dries007.tfc.objects.items.itemblock.ItemBlockTorchTFC;
+import net.dries007.tfc.objects.items.ItemBlockTFC;
+import net.dries007.tfc.objects.items.ItemBlockTorchTFC;
 
 import static net.dries007.tfc.Constants.MOD_ID;
 import static net.dries007.tfc.objects.CreativeTabsTFC.*;
