+++ src/main/java/net/dries007/tfc/objects/items/ItemMisc.java
@@ -1,39 +1,43 @@
 /*
  * Work under Copyright. Licensed under the EUPL.
  * See the project README.md and LICENSE.txt for more information.
+ *
  */
 
-package net.dries007.tfc.objects.items.wood;
+package net.dries007.tfc.objects.items;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 
-import net.minecraft.block.Block;
 import net.minecraft.item.ItemStack;
 
 import mcp.MethodsReturnNonnullByDefault;
 import net.dries007.tfc.objects.Size;
 import net.dries007.tfc.objects.Weight;
-import net.dries007.tfc.objects.items.ItemBlockTFC;
-import net.dries007.tfc.util.IPlacableItem;
+import net.dries007.tfc.util.IItemSize;
 
 @MethodsReturnNonnullByDefault
 @ParametersAreNonnullByDefault
-public class ItemLogTFC extends ItemBlockTFC implements IPlacableItem
+public class ItemMisc extends ItemTFC implements IItemSize
 {
-    public ItemLogTFC(Block block)
+    private final Size size;
+    private final Weight weight;
+
+    public ItemMisc(Size size, Weight weight)
     {
-        super(block);
+        super();
+        this.size = size;
+        this.weight = weight;
     }
 
     @Override
     public Size getSize(ItemStack stack)
     {
-        return Size.NORMAL;
+        return size;
     }
 
     @Override
     public Weight getWeight(ItemStack stack)
     {
-        return Weight.MEDIUM;
+        return weight;
     }
 }
